Refactor MulticallClient to be configurable
