+++ packages/backend/src/peripherals/ethereum/calls/BalanceCall.test.ts
@@ -3,7 +3,7 @@ import { AssetId, Bytes, EthereumAddress } from '@l2beat/shared-pure'
 import { expect } from 'earl'
 import { utils } from 'ethers'
 
-import { MULTICALL_V1_ADDRESS } from '../MulticallClient'
+import { MULTICALL_V1_ADDRESS } from '../MulticallConfig'
 import { BalanceCall } from './BalanceCall'
 
 describe('BalanceCall', () => {
