+++ packages/backend/src/peripherals/ethereum/types.ts
@@ -10,3 +10,21 @@ export interface CallParameters {
   value?: bigint
   data?: Bytes
 }
+
+export interface MulticallConfigEntry {
+  blockNumber: number
+  batchSize: number
+  address: EthereumAddress
+  encodeBatch: (requests: MulticallRequest[]) => Bytes
+  decodeBatch: (result: Bytes) => MulticallResponse[]
+}
+
+export interface MulticallRequest {
+  address: EthereumAddress
+  data: Bytes
+}
+
+export interface MulticallResponse {
+  success: boolean
+  data: Bytes
+}
